+++ source/core.py
@@ -214,6 +214,28 @@ class NewNVDAInstance:
 	directory: Optional[str] = None
 
 
+def computeRestartCLIArgs(removeArgsList: list[str] | None = None) -> list[str]:
+	"""Generate an equivalent list of CLI arguments from the values in globalVars.appArgs.
+	:param removeArgsList: A list of value to ignore when looking in globalVars.appArgs.
+	"""
+	from __main__ import parser  # import from nda.pyw
+
+	if not removeArgsList:
+		removeArgsList = []
+	args = []
+	for arg, val in globalVars.appArgs._get_kwargs():
+		if val == parser.get_default(arg):
+			continue
+		if arg in removeArgsList:
+			continue
+		flag = [a.option_strings[0] for a in parser._actions if a.dest == arg][0]
+		args.append(flag)
+		if isinstance(val, bool):
+			continue
+		args.append(f"{val}")
+	return args
+
+
 def restartUnsafely():
 	"""Start a new copy of NVDA immediately.
 	Used as a last resort, in the event of a serious error to immediately restart NVDA without running any
@@ -239,13 +261,16 @@ def restartUnsafely():
 			sys.argv.remove(paramToRemove)
 		except ValueError:
 			pass
+	restartCLIArgs = computeRestartCLIArgs(
+		removeArgsList=["easeOfAccess"],
+	)
 	options = []
 	if NVDAState.isRunningAsSource():
 		options.append(os.path.basename(sys.argv[0]))
 	_startNewInstance(
 		NewNVDAInstance(
 			sys.executable,
-			subprocess.list2cmdline(options + sys.argv[1:]),
+			subprocess.list2cmdline(options + restartCLIArgs),
 			globalVars.appDir,
 		),
 	)
@@ -260,15 +285,9 @@ def restart(disableAddons=False, debugLogging=False):
 		return
 	import subprocess
 
-	for paramToRemove in (
-		"--disable-addons",
-		"--debug-logging",
-		"--ease-of-access",
-	) + languageHandler.getLanguageCliArgs():
-		try:
-			sys.argv.remove(paramToRemove)
-		except ValueError:
-			pass
+	restartCLIArgs = computeRestartCLIArgs(
+		removeArgsList=["disableAddons", "debugLogging", "language", "easeOfAccess"],
+	)
 	options = []
 	if NVDAState.isRunningAsSource():
 		options.append(os.path.basename(sys.argv[0]))
@@ -280,7 +299,7 @@ def restart(disableAddons=False, debugLogging=False):
 	if not triggerNVDAExit(
 		NewNVDAInstance(
 			sys.executable,
-			subprocess.list2cmdline(options + sys.argv[1:]),
+			subprocess.list2cmdline(options + restartCLIArgs),
 			globalVars.appDir,
 		),
 	):